% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgenRbinom.R
\name{dgenRbinom}
\alias{dgenRbinom}
\title{Data generating process for biased binary outcome under treatment.}
\usage{
dgenRbinom(num.obs = 50, p = 0.3, q = 0.9, seed = 1)
}
\arguments{
\item{num.obs}{The number of observations.}

\item{p}{The probability of the outcome being 1 when treated.}

\item{q}{The probability of the outcome being 1 when treated.}

\item{seed}{The seed for randomization.}
}
\value{
A data frame of Y and D with \code{num.obs} total observations.
}
\description{
Data generating process for biased binary outcome under treatment.
}
\examples{
dgenRbinom(num.obs=50, p=0.3, q=0.9, seed=1)
}
