% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqnBB.R
\name{posterior.predictive.outcome}
\alias{posterior.predictive.outcome}
\title{Posterior predictive distribution of the outcome}
\usage{
posterior.predictive.outcome(Y, D, d, sigma0 = 1, lambda = 1)
}
\arguments{
\item{Y}{Outcome variable}

\item{D}{Treatment indicator variable}

\item{d}{New observation designation}

\item{sigma0}{Variance common to all disturbances}

\item{lambda}{The prior on \code{sigma0}}
}
\value{
Return the predictive posterior distribution of the outcome for a given treatment designation and existing data.
}
\description{
Return the parameters of the posterior distribution of the coefficients of the linear model.
Data (Y, D) and potential input d must be provided; sigma0 and lambda (the prior on sigma0)
are assumed to be known.
}
\examples{
df <- dgenRbinom()
posterior.predictive.outcome(df$Y, df$D,  d=0, sigma0=1, lambda=1)
posterior.predictive.outcome(df$Y, df$D,  d=1, sigma0=1, lambda=1)
}
