% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriorPredictiveOutcome.R
\name{posteriorPredictiveOutcome}
\alias{posteriorPredictiveOutcome}
\title{Posterior predictive distribution of the outcome}
\usage{
posteriorPredictiveOutcome(Y, D, d, sigma0 = 1, lambda = 1)
}
\arguments{
\item{Y}{Outcome variable}

\item{D}{Treatment indicator variable}

\item{d}{New observation designation}

\item{sigma0}{Variance common to all disturbances}

\item{lambda}{The prior on \code{sigma0}}
}
\value{
Return the predictive posterior distribution of the outcome for a given treatment designation and existing data.
}
\description{
Return the parameters of the posterior distribution of the coefficients of the linear model.
Data \code{(Y, D)} and potential input d must be provided; \code{sigma0} and \code{lambda} (the prior on \code{sigma0})
are assumed to be known.
}
\examples{
df <- dgenRbinom()
posteriorPredictiveOutcome(df$Y, df$D,  d=0, sigma0=1, lambda=1)
posteriorPredictiveOutcome(df$Y, df$D,  d=1, sigma0=1, lambda=1)
}
