% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriorTheta.R
\name{posteriorTheta}
\alias{posteriorTheta}
\title{Posterior distribution of model parameters}
\usage{
posteriorTheta(Y, D, sigma0 = 1, lambda = 1)
}
\arguments{
\item{Y}{Outcome variable}

\item{D}{Treatment indicator variable}

\item{sigma0}{Variance common to all disturbances}

\item{lambda}{The prior on \code{sigma0}}
}
\value{
Parameter estimates for the coefficients of the linear model.
}
\description{
Return the parameters of the posterior distribution of the coefficients of the linear model.
Data \code{(Y, D)} must be provided; \code{sigma0} and \code{lambda} (the prior on \code{sigma0}) are assumed to be known.
}
\examples{
df <- dgenRbinom()
posteriorTheta(df$Y, df$D, sigma0=1, lambda=1)
}
