% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriorThetaDensity.R
\name{posteriorThetaDensity}
\alias{posteriorThetaDensity}
\title{Density of the posterior distribution of model parameters at a given value of those model parameters}
\usage{
posteriorThetaDensity(theta, ptmap)
}
\arguments{
\item{theta}{specific values for the model parameters; this needs to be of length \code{dim(as.matrix(cbind(1,D)))[2]}}

\item{ptmap}{a \code{posteriorThetaMAP()} object}
}
\value{
the density of the posterior distribution of the model parameter estimates at the specific theta values.
}
\description{
Density of the posterior distribution of model parameters at a given value of those model parameters
}
\examples{
df <- dgenRbinom()
postThetaMAP <- posteriorThetaMAP(df$Y, df$D, sigma0=1, lambda=1)
theta <- c(0.3,0.9)
posteriorThetaDensity(theta, postThetaMAP)
}
