% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriorThetaEval.R
\name{posteriorThetaEval}
\alias{posteriorThetaEval}
\title{EquationEE}
\usage{
posteriorThetaEval(Y, D, theta, sigma0 = 1, lambda = 1)
}
\arguments{
\item{Y}{Outcome variable}

\item{D}{Treatment indicator variable}

\item{theta}{Specific values for the model parameters}

\item{sigma0}{Variance common to all disturbances}

\item{lambda}{The prior on \code{sigma0}}
}
\value{
The density of the posterior distribution of the model parameter estimates at the specified parameter values theta.
}
\description{
EquationEE
}
\examples{
df <- dgenRbinom()
thetaValues <- c(0.3,.9)
posteriorThetaEval(df$Y, df$D, thetaValues, sigma0=1, lambda=1)
}
