% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppDensity.R
\name{ppDensity}
\alias{ppDensity}
\title{Density of the posterior predictive distribution at the given outcome value y}
\usage{
ppDensity(y, pp)
}
\arguments{
\item{y}{the value taken by the new output}

\item{pp}{A \code{posteriorPredictive()} object}
}
\value{
the density of the posterior predictive distribution of the outcome taking value \code{y}
}
\description{
Return the density for y taking a certain value given the DGP. Requires
output from a \code{posteriorPredictiveOutcome()} object and the outcome value \code{y} at which
the density is measured.
}
\examples{
df <- dgenRbinom()
postPred <- posteriorPredictiveMAP(df$Y, df$D,  d=0, sigma0=1, lambda=1)
ppDensity(y=1, postPred)
}
