% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqnCC.R
\name{ppo.density}
\alias{ppo.density}
\title{Density of posterior predictive distribution at the given outcome value y}
\usage{
ppo.density(y, ppo)
}
\arguments{
\item{y}{Outcome variable value}

\item{ppo}{A posterior.predictive.outcome() object}
}
\value{
The density of the posterior predictive distribution of the outcome taking value y
}
\description{
Return the density for y taking a certain value given the DGP. Requires
output from posterior.predictive.outcome and the outcome value at which
the density is measured.
}
\examples{
df <- dgenRbinom()
ppoD0 <- posterior.predictive.outcome(df$Y, df$D,  d=0, sigma0=1, lambda=1)
ppo.density(y=1, ppoD0)
}
