% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppoDensity.R
\name{ppoDensity}
\alias{ppoDensity}
\title{Density of posterior predictive distribution at the given outcome value y}
\usage{
ppoDensity(y, ppo)
}
\arguments{
\item{y}{Outcome variable value}

\item{ppo}{A \code{posteriorPredictiveOutcome()} object}
}
\value{
The density of the posterior predictive distribution of the outcome taking value \code{y}
}
\description{
Return the density for y taking a certain value given the DGP. Requires
output from a \code{posteriorPredictiveOutcome()} object and the outcome value \code{y} at which
the density is measured.
}
\examples{
df <- dgenRbinom()
ppoD <- posteriorPredictiveOutcome(df$Y, df$D,  d=0, sigma0=1, lambda=1)
ppoDensity(y=1, ppoD)
}
