% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqnDD.R
\name{pt.density}
\alias{pt.density}
\title{Density of the posterior distribution of model parameters at a given value of those model parameters.}
\usage{
pt.density(theta, pt)
}
\arguments{
\item{theta}{A vector of model coefficients}

\item{pt}{The posterior distribution of the model parameters.}
}
\value{
The density of the posterior distribution of the model parameter estimates at the specific theta values.
}
\description{
Density of the posterior distribution of model parameters at a given value of those model parameters.
}
\examples{
df <- dgenRbinom()
post <- posterior.theta(df$Y, df$D, sigma0=1, lambda=1)
thetaValues <- c(0.3,0.9)
pt.density(thetaValues, post)
}
