% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptauDensity.R
\name{ptauDensity}
\alias{ptauDensity}
\title{Density of the posterior distribution of model parameters when the intercept is fixed at its MAP value and the treatment effect is given}
\usage{
ptauDensity(tau.value, pt)
}
\arguments{
\item{tau.value}{a specific value for the treatment effect parameter}

\item{pt}{a \code{posteriorTheta()} object}
}
\value{
of the posterior distribution of model parameters when the intercept is fixed at its MAP value and the treatment effect is given
}
\description{
Density of the posterior distribution of model parameters when the intercept is fixed at its MAP value and the treatment effect is given
}
\examples{
df <- dgenRbinom()
post <- posteriorTheta(df$Y, df$D, sigma0=1, lambda=1)
tau <- 0.1
posteriorTauDensity(tau, post)
}
