% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptDensity.R
\name{ptDensity}
\alias{ptDensity}
\title{Density of the posterior distribution of model parameters at a given value of those model parameters.}
\usage{
ptDensity(theta, pt)
}
\arguments{
\item{theta}{specific values for the model parameters; this needs to be of length \code{dim(as.matrix(cbind(1,D)))[2]}}

\item{pt}{a \code{posteriorTheta()} object}
}
\value{
the density of the posterior distribution of the model parameter estimates at the specific theta values.
}
\description{
Density of the posterior distribution of model parameters at a given value of those model parameters.
}
\examples{
df <- dgenRbinom()
post <- posteriorThetaMAP(df$Y, df$D, sigma0=1, lambda=1)
thetaValues <- c(0.3,0.9)
ptDensity(thetaValues, post)
}
