% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgenSimpleLinear.R
\name{dgenSimpleLinear}
\alias{dgenSimpleLinear}
\title{Data generating process for a simple linear regression}
\usage{
dgenSimpleLinear(num.obs = 50, intercept = 2, treatment.effect = -4, seed = 1)
}
\arguments{
\item{num.obs}{The number of observations.}

\item{intercept}{The y-intercept of the linear model.}

\item{treatment.effect}{The coefficient on the treatment indicator.}

\item{seed}{The seed for randomization.}
}
\value{
A data frame of Y and D with \code{num.obs} total observations.
}
\description{
Data generating process for a simple linear regression
}
\examples{
dgenSimpleLinear(num.obs=50, intercept=2, treatment.effect=-4, seed=1)
}
