% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriorTauDensity.R
\name{posteriorTauDensity}
\alias{posteriorTauDensity}
\title{Density of the posterior distribution of model parameters when the intercept is fixed at its MAP value and the treatment effect is given}
\usage{
posteriorTauDensity(Y, D, tau.value, sigma0 = 1, lambda = 1)
}
\arguments{
\item{Y}{an N-vector of outcomes}

\item{D}{an N-vector of treatment assignments}

\item{tau.value}{a specific value for the treatment effect parameter}

\item{sigma0}{prior for the standard deviation of the disturbance term in the linear model}

\item{lambda}{hyperprior for the standard deviation of the disturbance term in the linear model (\code{sigma0})}
}
\value{
of the posterior distribution of model parameters when the intercept is fixed at its MAP value and the treatment effect is given
}
\description{
Density of the posterior distribution of model parameters when the intercept is fixed at its MAP value and the treatment effect is given
}
\examples{
df <- dgenRbinom()
post <- posteriorThetaMAP(df$Y, df$D, sigma0=1, lambda=1)
tau <- 0.1
posteriorTauDensity(tau, post)
}
