% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriorThetaDensity.R
\name{posteriorThetaDensity}
\alias{posteriorThetaDensity}
\title{Density of the posterior distribution of the model parameter estimates at the specified values}
\usage{
posteriorThetaDensity(Y, D, theta, sigma0 = 1, lambda = 1)
}
\arguments{
\item{Y}{an N-vector of outcomes}

\item{D}{an N-vector of treatment assignments}

\item{theta}{specific values for the model parameters; this needs to be of length \code{dim(as.matrix(cbind(1,D)))[2]}}

\item{sigma0}{prior for the standard deviation of the disturbance term in the linear model}

\item{lambda}{hyperprior for the standard deviation of the disturbance term in the linear model (\code{sigma0})}
}
\value{
the density of the posterior distribution of the model parameter estimates at the specified values theta
}
\description{
Density of the posterior distribution of the model parameter estimates at the specified values
}
\examples{
df <- dgenRbinom(num.obs=50, p=0.3, q=0.9, seed=1) # Simulate some data
thetaValues <- c(0.3,.9) # Specify the model parameters (an intercept and a treatment effect)
posteriorThetaDensity(df$Y, df$D, thetaValues, sigma0=1, lambda=1)
}
